+++ index.html
@@ -229,7 +229,7 @@
   </template>
 
 
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
 </html>
